+++ MDFS5/src/main/java/edu/tamu/lenss/MDFS/Model/MDFSRsockBlockForFileCreate.java
@@ -11,16 +11,16 @@ public class MDFSRsockBlockForFileCreate implements Serializable {
 	public byte[] fileFrag;			//the byteArray that contains filefrag
 	public String fileID;				//fileiD
 	public long entireFileSize;		//size of the entire file
-	public byte n2;					//n2
-	public byte k2;					//k2
-	public byte blockIdx;			//block number of this file
-	public byte fragmentIdx;		//fragment number of this block
+	public int n2;					//n2
+	public int k2;					//k2
+	public int blockIdx;			//block number of this file
+	public int fragmentIdx;		//fragment number of this block
 	public String fileCreatorGUID;  //the guid/node which created the file
 	public String uniqueReqID;		//unique blockRetrieveReqUUID using which this file metadata is identified
 	public boolean isGlobal;        //is this file globally accessible or nah
 
 
-	public MDFSRsockBlockForFileCreate(String filename, String filePathMDFS, byte[] filefrag, String fileID, long entirefilesize, byte n2, byte k2, byte blockidx, byte fragmentidx, String filecreatorguid, String uniquereqid, boolean isglobal){
+	public MDFSRsockBlockForFileCreate(String filename, String filePathMDFS, byte[] filefrag, String fileID, long entirefilesize, int n2, int k2, int blockidx, int fragmentidx, String filecreatorguid, String uniquereqid, boolean isglobal){
 		this.fileName = filename;
 		this.filePathMDFS = filePathMDFS;
 		this.fileFrag = filefrag;
